+++ CHANGES.rst
@@ -20,7 +20,8 @@ Imviz
 - Added the ability to load DS9 region files (``.reg``) using the ``IMPORT DATA``
   button. However, this only works after loading at least one image into Imviz. [#2201]
 
-- Added support for new ``CircularAnnulusROI`` subset from glue. [#2201]
+- Added support for new ``CircularAnnulusROI`` subset from glue, including
+  a new draw tool. [#2201, #2240]
 
 Mosviz
 ^^^^^^